+++ packages/dumb/src/components/trends/cbg/cbg-median-animated/cbg-median.css
