+++ packages/dumb/src/components/trends/cbg/cbg-slice/cbg-slices-container.tsx
@@ -29,8 +29,8 @@ import React, { type FunctionComponent, memo, useMemo } from 'react'
 
 import { type BgBounds } from 'medical-domain'
 import { type CbgSlicesContainerData } from '../../../../models/cbg-slices-container-data.model'
-import { CbgSliceAnimated } from './cbg-slice-animated'
-import { CbgMedianAnimated } from '../cbg-median-animated/cbg-median-animated'
+import { CbgSliceSegment } from './cbg-slice-segment'
+import { CbgMedian } from '../cbg-median-animated/cbg-median'
 import { formatCbgs } from './cbg-slices-container.util'
 import { type ScaleFunction } from '../../../../models/scale-function.model'
 import { useTrendsContext } from '../../../../provider/trends.provider'
@@ -65,7 +65,7 @@ const CbgSlicesContainer: FunctionComponent<CbgSlicesContainerProps> = (props) =
     <>
       {cbgs.map(cbg => (
         <g key={cbg.id} data-testid="cbg-slice-segments">
-          <CbgSliceAnimated
+          <CbgSliceSegment
             datum={cbg}
             sliceWidth={sliceWidth}
             tooltipLeftThreshold={tooltipLeftThreshold}
@@ -74,7 +74,7 @@ const CbgSlicesContainer: FunctionComponent<CbgSlicesContainerProps> = (props) =
             yScale={yScale}
           />
           {displayFlags.cbgMedianEnabled &&
-            <CbgMedianAnimated
+            <CbgMedian
               bgBounds={bgBounds}
               median={cbg.median}
               msX={cbg.msX}
